Tests: adjusted limit_rate tests for leaky bucket implementation.
Previously, stream limit_rate tests were quite dependant on the particular
limit_rate implementation.  Relaxed testing so it would also succeed
with the upcoming leaky bucket implementation, and described high-level
requirements in the comment instead of the behaviour of the old
implementation.

Similarly, comment in proxy_limit_rate.t was replaced with a high-level
description instead of an (incorrect) description of the behaviour of
the old implementation.