Tests: improved h2_priority.t.
Tests in h2_priority.t do not tolerate extra data being split over multiple
HTTP/2 frames, yet this can happen if writing to the socket buffer blocks
and then a small amount of data (1 byte) remains in the HTTP/2 connection
queue due to flow control: when the window is opened again, this 1 byte
will be sent in a separate frame, and the remaining bytes will be read and
sent afterwards.

Failures due to this issue were observed during testing on OpenBSD, and
might be easily reproduced on other platforms with small enough socket
buffers.

Fix is to use output_buffers large enough to read the full response, so the
remaining bytes are always sent in a single HTTP/2 frame.