Tests: fixed ssl_certificates.t with LibreSSL client.
When Net::SSLeay is built with LibreSSL, it is not possible to control
signature algorithms sent to the server, which results in incorrect
certificate selection with TLSv1.3.  Notably, with LibreSSL on server
an ECDSA certificate is always used, and with OpenSSL an RSA certificate.

Following 2005:00307a7f3cad, TODO in a test expecting an RSA certificate is
no longer used for LibreSSL 4.0.0 and above.  This works with OpenSSL on
the server, but breaks when LibreSSL is used on the server.

Fix is to mark the test as TODO when LibreSSL (any version) is used both
on the server and by the Net::SSLeay library.