Tests: removed bytes sent tests on rejected connection upgrades.
In proxy_upgrade.t and ssl_proxy_upgrade.t, $body_bytes_sent was expected
to be 0 if connection upgrade was not requested by the client, yet the
upstream server tried to return 101 (Switching Protocols).  This is,
however, not the only possible valid outcome.  In particular, if such
a response is considered to be an error, and 502 (Bad Gateway) is generated,
$body_bytes_sent won't be 0, breaking the test.  As such, these tests
were removed.