Tests: adjusted delays in ssl_session_ticket_key.t.
If session ticket keys are not shared among worker processes, waiting
for 0.5 seconds before trying to establish another session won't change
anything (if the connection will end up in the different worker process,
there will be an error).

At the same time, the test was seen unexpectedly failing with Valgrind,
due to slow connection handling.  Since session ticket key rotation uses
time in seconds, slightly more than 1 second between requests is enough
for the test to fail if the first request happens at the end of a second.
Removing the unneeded sleep improves things, but not enough to fully
resolve the issue.  As such, session timeout was set to 3 seconds.

Wait time between tests adjusted accordingly, to slightly more than 4
seconds, to ensure that the key will be changed even if the first request
happens at the start of a second.